bad request validaiton

+++ HR.LeaveManagement.Application/Features/LeaveType/Commands/DeleteLeaveType/DeleteLeaveTypeCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using HR.LeaveManagement.Application.Contracts.Persistence;
+using HR.LeaveManagement.Application.Exceptions;
 using MediatR;
 
 namespace HR.LeaveManagement.Application.Features.LeaveType.Commands.DeleteLeaveType;
@@ -14,6 +15,9 @@ public class DeleteLeaveTypeCommandHandler : IRequestHandler<DeleteLeaveTypeComm
     public async Task<Unit> Handle(DeleteLeaveTypeCommand request, CancellationToken cancellationToken)
     {
         var leaveTypeDelete = await _leaveTypeRepository.GetByIdAsync(request.Id);
+
+        if(leaveTypeDelete == null) 
+            throw new NotFoundException(nameof(LeaveType), request.Id);
 
         await _leaveTypeRepository.DeleteAsync(leaveTypeDelete);
